stable rust

+++ src/modes/dungeon/systems/render/gui/mod.rs
@@ -23,14 +23,14 @@ impl Plugin for GUIPlugin {
             GameStage::Render,
             ConditionSet::new()
                 .run_if(run_in_state(TurnState::ShowInventory))
-                .with_system(menus::show_inventory::<{ InventoryMenu::Main }>)
+                .with_system(menus::show_inventory::<{ InventoryMenu::Main as u8 }>)
                 .into(),
         )
         .add_system_set_to_stage(
             GameStage::Render,
             ConditionSet::new()
                 .run_if(run_in_state(TurnState::ShowDropMenu))
-                .with_system(menus::show_inventory::<{ InventoryMenu::Drop }>)
+                .with_system(menus::show_inventory::<{ InventoryMenu::Drop as u8 }>)
                 .into(),
         );
     }
